Add some common,now support multiple framework!
