+++ BinSerializer/BinSerializer.csproj
@@ -1,80 +1,49 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<Project ToolsVersion="14.0" DefaultTargets="Build" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
-  <Import Project="$(MSBuildExtensionsPath)\$(MSBuildToolsVersion)\Microsoft.Common.props" Condition="Exists('$(MSBuildExtensionsPath)\$(MSBuildToolsVersion)\Microsoft.Common.props')" />
-  <PropertyGroup>
-    <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
-    <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
-    <ProjectGuid>{15EDB439-1FED-4BC8-B27F-4A3B867A3F95}</ProjectGuid>
-    <OutputType>Library</OutputType>
-    <AppDesignerFolder>Properties</AppDesignerFolder>
-    <RootNamespace>ThirtyNineEighty.BinarySerializer</RootNamespace>
-    <AssemblyName>BinSerializer</AssemblyName>
-    <TargetFrameworkVersion>v4.5</TargetFrameworkVersion>
-    <FileAlignment>512</FileAlignment>
-    <TargetFrameworkProfile />
-    <CopyNuGetImplementations>false</CopyNuGetImplementations>
-  </PropertyGroup>
-  <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Debug|AnyCPU' ">
-    <DebugSymbols>true</DebugSymbols>
-    <DebugType>full</DebugType>
-    <Optimize>false</Optimize>
-    <OutputPath>bin\Debug\</OutputPath>
-    <DefineConstants>DEBUG;TRACE</DefineConstants>
-    <ErrorReport>prompt</ErrorReport>
-    <WarningLevel>4</WarningLevel>
-    <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
-  </PropertyGroup>
-  <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Release|AnyCPU' ">
-    <DebugType>pdbonly</DebugType>
-    <Optimize>true</Optimize>
-    <OutputPath>bin\Release\</OutputPath>
-    <DefineConstants>TRACE</DefineConstants>
-    <ErrorReport>prompt</ErrorReport>
-    <WarningLevel>4</WarningLevel>
-    <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
-  </PropertyGroup>
-  <PropertyGroup>
-    <SignAssembly>true</SignAssembly>
-  </PropertyGroup>
-  <PropertyGroup>
-    <AssemblyOriginatorKeyFile>BinSerializerStrongName.snk</AssemblyOriginatorKeyFile>
-  </PropertyGroup>
-  <ItemGroup>
-    <Reference Include="System" />
-    <Reference Include="System.Core" />
-    <Reference Include="Microsoft.CSharp" />
-  </ItemGroup>
-  <ItemGroup>
-    <Compile Include="BinField.cs" />
-    <Compile Include="BinSerializer.cs" />
-    <Compile Include="BSDebug.cs" />
-    <Compile Include="BinFieldAttribute.cs" />
-    <Compile Include="IBinSerializable.cs" />
-    <Compile Include="MethodAdapter.cs" />
-    <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="RefReaderWatcher.cs" />
-    <Compile Include="RefWriterWatcher.cs" />
-    <Compile Include="SerializerBuilder.cs" />
-    <Compile Include="Types\ArraySerializerTypeInfo.cs" />
-    <Compile Include="Types\BinTypeDescription.cs" />
-    <Compile Include="Types\BinTypeProcess.cs" />
-    <Compile Include="Types\BinTypeVersion.cs" />
-    <Compile Include="Types\GenericSerializerTypeInfo.cs" />
-    <Compile Include="Types\SerializerTypeInfo.cs" />
-    <Compile Include="BinStreamExtensions.cs" />
-    <Compile Include="BinTypeAttribute.cs" />
-    <Compile Include="Types\SerializerTypes.cs" />
-    <Compile Include="BinTypeExtensions.cs" />
-  </ItemGroup>
-  <ItemGroup>
-    <None Include="BinSerializerStrongName.snk" />
-  </ItemGroup>
-  <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
-  <!-- To modify your build process, add your task inside one of the targets below and uncomment it. 
-       Other similar extension points exist, see Microsoft.Common.targets.
-  <Target Name="BeforeBuild">
-  </Target>
-  <Target Name="AfterBuild">
-  </Target>
-  -->
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+	<PropertyGroup>
+		<TargetFrameworks>netcoreapp3.1;net45;net452;net462;netstandard2.1;net50</TargetFrameworks>
+		<SignAssembly>true</SignAssembly>
+		<AssemblyOriginatorKeyFile>BinSerializerStrongName.snk</AssemblyOriginatorKeyFile>
+		<PackageId>ThirtyNineEighty.BinarySerializer</PackageId>
+		<Authors>ThirtyNineEighty.BinarySerializer</Authors>
+		<Product>BinSerializer</Product>
+		<GeneratePackageOnBuild>true</GeneratePackageOnBuild>
+		<Company>ThirtyNineEighty</Company>
+		<Copyright>Copyright © ThirtyNineEighty 2016</Copyright>
+		<Description>Binary serializer</Description>
+		<Version>2.2.0.0</Version>
+		<PackageProjectUrl>https://github.com/Nirklav/BinSerializer</PackageProjectUrl>
+		<RepositoryUrl>https://github.com/Nirklav/BinSerializer</RepositoryUrl>
+		<RepositoryType>git</RepositoryType>
+	</PropertyGroup>
+
+	<PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|AnyCPU'">
+	  <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
+	</PropertyGroup>
+
+	<PropertyGroup >
+	  <DocumentationFile>D:\Users\WXF\Source\Repos\WitchieBinSerializer\BinSerializer\BinSerializer.xml</DocumentationFile>
+	</PropertyGroup>
+
+	<ItemGroup Condition="'$(TargetFramework)' == 'netcoreapp3.1'">
+	  <PackageReference Include="System.Security.Permissions">
+	    <Version>4.7.0</Version>
+	  </PackageReference>
+	</ItemGroup>
+	<ItemGroup Condition="'$(TargetFramework)' == 'net50'">
+		<PackageReference Include="System.Security.Permissions">
+			<Version>5.0.0-rc.2.20475.5</Version>
+		</PackageReference>
+	</ItemGroup>
+	<ItemGroup Condition="'$(TargetFramework)' == 'net462'">
+	  <PackageReference Include="System.Security.Permissions">
+	    <Version>4.7.0</Version>
+	  </PackageReference>
+	</ItemGroup>
+
+	<ItemGroup Condition="'$(TargetFramework)' == 'netstandard2.1'">
+	  <PackageReference Include="System.Security.Permissions">
+	    <Version>4.7.0</Version>
+	  </PackageReference>
+	</ItemGroup>
 </Project>
