+++ BinSerializer/BinStream/StreamExtensionKind.cs
@@ -1,0 +1,9 @@
+﻿namespace ThirtyNineEighty.BinarySerializer
+{
+    public enum StreamExtensionKind
+  {
+    Write,
+    Read,
+    Skip
+  }
+}
