+++ BinSerializer/BinType/BinTypeAttribute.cs
