+++ BinSerializer/BinType/TypeExtensionKind.cs
@@ -1,0 +1,8 @@
+﻿namespace ThirtyNineEighty.BinarySerializer
+{
+    enum TypeExtensionKind
+  {
+    Write,
+    Read
+  }
+}
