+++ BinSerializer/DeserializationInfo.cs
@@ -1,0 +1,12 @@
+﻿namespace ThirtyNineEighty.BinarySerializer
+{
+    public struct DeserializationInfo
+  {
+    public readonly int Version;
+
+    public DeserializationInfo(int version)
+    {
+      Version = version;
+    }
+  }
+}
