+++ BinSerializer/Properties/AssemblyInfo.cs
@@ -2,37 +2,10 @@
 using System.Runtime.InteropServices;
 using System.Security;
 
-// General Information about an assembly is controlled through the following 
-// set of attributes. Change these attribute values to modify the information
-// associated with an assembly.
-[assembly: AssemblyTitle("BinSerializer")]
-[assembly: AssemblyDescription("Binary serializer")]
-[assembly: AssemblyConfiguration("")]
-[assembly: AssemblyCompany("ThirtyNineEighty")]
-[assembly: AssemblyProduct("BinSerializer")]
-[assembly: AssemblyCopyright("Copyright © ThirtyNineEighty 2016")]
-[assembly: AssemblyTrademark("")]
-[assembly: AssemblyCulture("")]
-
-// Setting ComVisible to false makes the types in this assembly not visible 
-// to COM components.  If you need to access a type in this assembly from 
-// COM, set the ComVisible attribute to true on that type.
 [assembly: ComVisible(false)]
 
 // The following GUID is for the ID of the typelib if this project is exposed to COM
 [assembly: Guid("15edb439-1fed-4bc8-b27f-4a3b867a3f95")]
 
-// Version information for an assembly consists of the following four values:
-//
-//      Major Version
-//      Minor Version 
-//      Build Number
-//      Revision
-//
-// You can specify all the values or you can default the Build and Revision Numbers 
-// by using the '*' as shown below:
-// [assembly: AssemblyVersion("1.0.*")]
-[assembly: AssemblyVersion("2.2.0.0")]
-[assembly: AssemblyFileVersion("2.2.0.0")]
 
 [assembly: AllowPartiallyTrustedCallers]
