+++ BinSerializer/SerializationInfo.cs
@@ -1,0 +1,6 @@
+﻿namespace ThirtyNineEighty.BinarySerializer
+{
+    public struct SerializationInfo
+  {
+  }
+}
